LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-affects-child-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/plugins/resources/mock-plugin.pl because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
This tests that a plugin-types directive on a top-level page should be inherited by a PluginDocument embedded in the page. This test passes if a console error says that the plugin was blocked.
